Parse out vertices and normals
We explictly expand element definitions into slices of vertices and
normals that we pass back to the caller. This is pretty useful in the
case of flat-shaded models where vertices with the same position won't
share normals, and it's slightly easier to deal with than a more general
solution for the moment.